+++ src/migrations/Install.php
@@ -494,6 +494,8 @@ class Install extends Migration
             'hasVariants' => $this->boolean(),
             'hasVariantTitleField' => $this->boolean(),
             'titleFormat' => $this->string()->notNull(),
+            'titleLabel' => $this->string()->defaultValue('Title'),
+            'variantTitleLabel' => $this->string()->defaultValue('Title'),
             'skuFormat' => $this->string(),
             'descriptionFormat' => $this->string(),
             'dateCreated' => $this->dateTime()->notNull(),