+++ src/services/Customers.php
@@ -713,7 +713,8 @@ class Customers extends Component
 
         Craft::$app->getView()->registerAssetBundle(CommerceCpAsset::class);
         return Craft::$app->getView()->renderTemplate('commerce/customers/_editUserTab', [
-            'customer' => $customer
+            'customer' => $customer,
+            'userId' => $context['user']->id,
         ]);
     }
 }