+++ src/services/Emails.php
@@ -317,11 +317,13 @@ class Emails extends Component
      * @param Email $email
      * @param Order $order
      * @param OrderHistory $orderHistory
+     * @param array $orderData Since the order may have changed by the time the email sends.
      * @return bool $result
      * @throws Exception
      * @throws Throwable
+     * @throws \yii\base\InvalidConfigException
      */
-    public function sendEmail($email, $order, $orderHistory = null): bool
+    public function sendEmail($email, $order, $orderHistory = null, $orderData = null): bool
     {
         if (!$email->enabled) {
             return false;
@@ -342,7 +344,7 @@ class Emails extends Component
         }
 
         //sending emails
-        $renderVariables = compact('order', 'orderHistory', 'option');
+        $renderVariables = compact('order', 'orderHistory', 'option', 'orderData');
 
         $mailer = Craft::$app->getMailer();
         /** @var Message $newEmail */
@@ -671,7 +673,8 @@ class Emails extends Component
                 'craftEmail' => $newEmail,
                 'commerceEmail' => $email,
                 'order' => $order,
-                'orderHistory' => $orderHistory
+                'orderHistory' => $orderHistory,
+                'orderData' => $orderData
             ]);
             $this->trigger(self::EVENT_BEFORE_SEND_MAIL, $event);
 
@@ -725,7 +728,8 @@ class Emails extends Component
                 'craftEmail' => $newEmail,
                 'commerceEmail' => $email,
                 'order' => $order,
-                'orderHistory' => $orderHistory
+                'orderHistory' => $orderHistory,
+                'orderData' => $orderData
             ]));
         }